Improve support for git patches
When git-2.4 formats a patch containing an unaltered empty line it
puts an empty line ('\n') in the output. unidiff expects such lines
to have a leading space (' \n') and throws an exception when presented
with such patches.

This patch fixes parser to cope with this input and updates the git
test file so that it contains an empty line.