Optimize for large hunk sizes.
The processing for each hunk is currently O(N^2) causing patch
processing times to explode for patches that introduce a new file.
For example, on a 4GHz AMD CPU it takes over a minute to handle a
13,000 line patch file.

When the O(N^2) is replacd with an O(N) algorithm the same file can
be processed in half a second!